move verb functions to verb controller
